+++ packages/dsw-mailer/dsw/mailer/model.py
@@ -49,9 +49,15 @@ class TemplateDescriptorPart:
 
 class TemplateDescriptor:
 
-    def __init__(self, message_id: str, subject: str):
+    def __init__(self, message_id: str, subject: str, language: str,
+                 importance: str, sensitivity: Optional[str],
+                 priority: Optional[str]):
         self.id = message_id
         self.subject = subject
+        self.language = language
+        self.importance = importance
+        self.sensitivity = sensitivity
+        self.priority = priority
         self.parts = []  # type: list[TemplateDescriptorPart]
         self.modes = []  # type: list[str]
 
@@ -60,6 +66,10 @@ class TemplateDescriptor:
         result = TemplateDescriptor(
             message_id=data.get('id', ''),
             subject=data.get('subject', ''),
+            language=data.get('language', 'en'),
+            importance=data.get('importance', 'normal'),
+            sensitivity=data.get('sensitivity', None),
+            priority=data.get('priority', None),
         )
         result.parts = [TemplateDescriptorPart.load_from_file(d)
                         for d in data.get('parts', [])]
@@ -76,6 +86,8 @@ class MessageRequest:
         self.trigger = trigger
         self.ctx = ctx
         self.recipients = recipients
+        self.domain = None  # type: Optional[str]
+        self.client_url = ''  # type: str
 
     @staticmethod
     def load_from_file(data: dict) -> 'MessageRequest':
@@ -99,6 +111,13 @@ class MailMessage:
         self.html_body = None  # type: Optional[str]
         self.html_images = list()  # type: list[MailAttachment]
         self.attachments = list()  # type: list[MailAttachment]
+        self.msg_id = None  # type: Optional[str]
+        self.msg_domain = None  # type: Optional[str]
+        self.language = 'en'  # type: str
+        self.importance = 'normal'  # type: str
+        self.sensitivity = None  # type: Optional[str]
+        self.priority = None  # type: Optional[str]
+        self.client_url = ''  # type: str
 
 
 class MailAttachment: